requirements.txt
asgiref
aspose-words==25.5.0
babel==2.17.0
Django>=5.0
docxcompose==1.4.0
gunicorn
lxml
mysqlclient
Pillow
plum-dispatch==1.7.4
python-docx==1.1.2
setuptools
six==1.17.0
spire-doc==13.4.6
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata
whitenoise
PyJWT
requests
pandas
openpyxl